requirements.txt
Flask
python-dotenv
pandas
matplotlib
numpy
